requirements.txt
bottle==0.12.8
Mako==1.0.1
MarkupSafe==0.23
mongoengine==0.8.7
Pillow==2.7.0
pymongo==2.8
requests==2.5.1
